seporate model use for each chat

+++ src/app/components/sidebar/sidebar.component.ts
@@ -6,9 +6,14 @@ import { ModelType } from '../../types/model-type';
 import { getModelLabelByKey } from '../../helpers/get-model-label-by-key';
 import { formatTimestamp } from '../../helpers/format-timestamp';
 import { ChatMessage } from '../../models/chat.model';
-import { AppService } from '../../services/app.service';
 import { TUI_CONFIRM, TuiConfirmData } from '@taiga-ui/kit';
 import { ChatState } from '../../types/chat-state';
+
+const TuiConfirmText: TuiConfirmData = {
+  content: 'Это действие нельзя будет отменить',
+  yes: 'Да',
+  no: 'Нет',
+};
 
 @Component({
   selector: 'app-sidebar',
@@ -21,49 +26,31 @@ export class SidebarComponent {
   protected readonly ChatState = ChatState;
   
   constructor(
-    private readonly appServbice: AppService,
     public readonly chatService: ChatService,
     private readonly dialogService: TuiDialogService
   ) {}
 
   createNewChat(): void  {
-    this.chatService.activeChatId.set(null)
-    if (this.appServbice.isMobile()) this.appServbice.sidebarOpen.set(false)
+    this.chatService.navigateToChat(null)
   }
 
   openChat(chatId: string): void {
-    this.chatService.activeChatId.set(chatId)
-    if (this.appServbice.isMobile()) this.appServbice.sidebarOpen.set(false)
+    this.chatService.navigateToChat(chatId)
   }
 
   deleteChat(event: MouseEvent, chatId: string): void {
     event.stopPropagation();
 
-    const data: TuiConfirmData = {
-      content: 'Это действие нельзя будет отменить',
-      yes: 'Да',
-      no: 'Нет',
-    };
-
     this.dialogService
-      .open<boolean>(TUI_CONFIRM, {size: 's', label: 'Удалить чат?', data})
+      .open<boolean>(TUI_CONFIRM, {size: 's', label: 'Удалить чат?', data: TuiConfirmText})
       .subscribe((confirm) => {
-        if (confirm) {
-          this.chatService.deleteChat(chatId)
-          if (this.appServbice.isMobile()) this.appServbice.sidebarOpen.set(false)
-        }
+        if (confirm) this.chatService.deleteChat(chatId)
       });
   }
 
   openClearConfirmationModal(): void {
-    const data: TuiConfirmData = {
-      content: 'Это действие нельзя будет отменить',
-      yes: 'Да',
-      no: 'Нет',
-    };
-
     this.dialogService
-      .open<boolean>(TUI_CONFIRM, {size: 's', label: 'Очистить историю?', data})
+      .open<boolean>(TUI_CONFIRM, {size: 's', label: 'Очистить историю?', data: TuiConfirmText})
       .subscribe((confirm) => {
         if (confirm) this.chatService.deleteAllChats()
       });
